api/requirements.txt
pandas==1.3.3
numpy==1.23.5
mtcnn==0.1.1
keras_facenet==0.3.2
tensorflow==2.12
flask==2.3.2
pillow
scipy
waitress